4.21*: blogilistan laajennus, step10

+++ controllers/blogs.js
@@ -1,6 +1,7 @@
 const blogsRouter = require('express').Router()
 const Blog = require('../models/blog')
 const User = require('../models/user')
+const jwt = require('jsonwebtoken')
 
 blogsRouter.get('/', async (request, response) => {
   const blogs = await Blog.find({})
@@ -11,36 +12,64 @@ blogsRouter.get('/', async (request, response) => {
 blogsRouter.post('/', async (request, response, next) => {
   const body = request.body
 
-  // const user = await User.findById(body.userId)
-  // Temporary hack:
-  const users = await User.find({})
-  const user = users[0]
-
-  var blog = new Blog({
-    title: body.title,
-    author: body.author,
-    url: body.url,
-    likes: body.likes = body.likes === undefined ? 0 : body.likes,
-    user: user.id
-  })
-
   try {
-    const savedBlog = await blog.save()
-    user.blogs = user.blogs.concat(savedBlog._id)
-    await user.save()
-    response.json(savedBlog.toJSON())
+    const decodedToken = jwt.verify(request.token, process.env.SECRET)
+    if (!request.token || !decodedToken.id) {
+      return response.status(401).json({ error: 'Token missing or invalid' })
+    }
+
+    const user = await User.findById(decodedToken.id)
+
+    var blog = new Blog({
+      title: body.title,
+      author: body.author,
+      url: body.url,
+      likes: body.likes = body.likes === undefined ? 0 : body.likes,
+      user: user.id
+    })
+
+    try {
+      const savedBlog = await blog.save()
+      user.blogs = user.blogs.concat(savedBlog._id)
+      await user.save()
+      response.json(savedBlog.toJSON())
+    } catch (exception) {
+      next(exception)
+    }
+
   } catch (exception) {
     next(exception)
   }
 })
 
 blogsRouter.delete('/:id', async (request, response, next) => {
+
   try {
-    await Blog.findByIdAndRemove(request.params.id)
-    response.status(204).end()
+    const decodedToken = jwt.verify(request.token, process.env.SECRET)
+    if (!request.token || !decodedToken.id) {
+      return response.status(401).json({ error: 'Token missing' })
+    }
+    const user = await User.findById(decodedToken.id)
+    const blog = await Blog.findById(request.params.id)
+    console.log('blog', blog)
+    if (blog === undefined || blog === null) {
+      return response.status(400).json({ error: 'Invalid blog id' })
+    }
+
+    if (blog.user.toString() === user.id) {
+      try {
+        await Blog.findByIdAndRemove(request.params.id)
+        response.status(204).end()
+      } catch (exception) {
+        next(exception)
+      }
+    }
   } catch (exception) {
     next(exception)
   }
+
+
+
 })
 
 blogsRouter.put('/:id', async (request, response, next) => {
